third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-invalid-hsl.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Level 4: Parsing and serialization of colors using invalid HSL notation</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#the-hsl-notation">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#resolving-sRGB-values">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-sRGB-values">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("color", "hsl(none, none, none)");
test_invalid_value("color", "hsla(none, none, none, none)");
test_invalid_value("color", "hsl(none, 100%, 50%)");
test_invalid_value("color", "hsla(120, 100%, 50%, none)");
</script>
</body>
</html>
